requirements.txt
streamlit>=1.37.0
pymongo>=4.0.0
dnspython>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.0.0
requests>=2.28.0
